requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
django-cors-headers==2.2.0
django-crispy-forms==1.7.0
django-filter==1.1.0
django==2.0.2
djangorestframework-jwt==1.11.0
djangorestframework==3.7.7
pillow==5.0.0
psycopg2-binary==2.7.4
pyjwt==1.6.0              # via djangorestframework-jwt
pytz==2018.3              # via django
